+++ documentation/changelog.rst
@@ -10,6 +10,9 @@ v0.20.0 | April XX, 2024
 New features
 -------------   
 
+* Add command ``flexmeasures edit transfer-ownership`` to transfer the ownership of an asset and its children from one account to another[see `PR #983 <https://github.com/FlexMeasures/flexmeasures/pull/983>`_]
+
+
 Infrastructure / Support
 ----------------------